+++ .github/workflows/galaxy-importer.cfg
@@ -1,0 +1,9 @@
+[galaxy-importer]
+LOG_LEVEL_MAIN = INFO
+RUN_FLAKE8 = True
+RUN_ANSIBLE_DOC = True
+RUN_ANSIBLE_LINT = True
+RUN_ANSIBLE_TEST = False
+ANSIBLE_TEST_LOCAL_IMAGE = False
+LOCAL_IMAGE_DOCKER = False
+INFRA_OSD = False
