+++ ansible.cfg
@@ -1,4 +1,3 @@
 [defaults]
-collections_paths = ./ansible_collections:/fmc-ansible/ansible_collections:/usr/local/lib/python3.6/site-packages/ansible_collections/:/root/.ansible/collections/ansible_collections
+#collections_paths = ./ansible_collections:/fmc-ansible/ansible_collections:/usr/local/lib/python3.6/site-packages/ansible_collections/:/root/.ansible/collections/ansible_collections
 log_path = ./logs
-
