--- ansible_collections/cisco/fmcansible/README.md
@@ -1,3 +1,0 @@
-# Ansible Collection - cisco.fmcansible
-
-Documentation for the collection.
