+++ changelogs/fragments/0.4.0.yml
@@ -1,0 +1,3 @@
+major_changes:
+  - Migrated modules to Ansible collection `cisco.ftdansible`
+  - fixed linting issues and other issues raised by ansible-test
