+++ samples/fmc_configuration/user.yml
@@ -1,0 +1,7 @@
+- hosts: all
+  connection: httpapi
+  tasks:
+    - name: Create an network object for Cisco FMC
+      cisco.fmcansible.fmc_configuration:
+        operation: getAllDomain
+        register_as: cra1
