--- tox.ini
@@ -1,43 +1,0 @@
-# tox (https://tox.readthedocs.io/) is a tool for running tests
-# in multiple virtualenvs. This configuration file will run the
-# test suite on all supported python versions. To use it, "pip install tox"
-# and then run "tox" from this directory.
-
-[tox]
-envlist = py{27,35,36,37}{,-integration}
-skipsdist = True
-
-[testenv]
-deps =
-    -r{toxinidir}/requirements.txt
-    !integration: -r{toxinidir}/test/requirements.txt
-    integration: junit_xml
-
-# Following line allow tox inherit predefined PYTHONPATH env var.
-# set PYTHONPATH=<some path> can be used instead as well
-passenv = PYTHONPATH
-setenv =
-    integration: ANSIBLE_STDOUT_CALLBACK = junit
-    integration: JUNIT_OUTPUT_DIR = {env:REPORTS_DIR}
-
-commands =
-    python --version
-# Run unit tests by default
-    !integration: pytest --junitxml=.tox/junit-{envname}.xml test/unit
-# Run integration test if in intgration env
-    integration: ansible-playbook {posargs}
-
-[flake8]
-ignore = E402
-max-line-length = 120
-exclude =
-    .git,
-    .tox,
-    .pytest_cache,
-    venv,
-    docs/dist,
-    docs/static
-    docs/templates,
-    inventory,
-    samples,
-    __pycache__
